Refined automate_notebook.py for automation of execution of notebooks; Removed WIP and experimental builds of automate_notebook.py
